Second set of examples - Only example 3 will match
Elixir doesn’t allow multiple references of the same variable when
matching unless the values on the right side are the same. For example,
[a, a, a] = [1, 1, 1] will match; so will [a, b, a] = [1, 2, 1] since
the variable a is always matching the value 1.

Elixir will fail the entire expression if any values do not match—
which is why the variable b in [a, b, a] = [1, 2, 3] will not contain a
value after the expression is evaluated.